feat: rename kubernetes_cleanup_namespaces -> kubernetes_cleanup_review_apps
