+++ Makefile
@@ -1,6 +1,6 @@
 test:
 	pytest src/kubernetes_cleanup_jobs
-	pytest src/kubernetes_cleanup_namespaces
+	pytest src/kubernetes_cleanup_review_apps
 	pytest src/kubernetes_cleanup_pods
 	pytest src/kubernetes_configmap_jwt_scan
 	pytest src/lib
