+++ src/kubernetes_cleanup_review_apps/cleanup.py
@@ -1,0 +1,13 @@
+import argparse
+import logging
+import os
+
+import migrate_config_secrets.context
+
+from kubernetes_cleanup_review_apps.namespaces import (
+  cleanup_namespaces
+)
+
+if __name__ == "__main__":
+
+  cleanup_namespaces()
